Remove current values from Circuit struct
Those values are not being used (they are only passed in CircuitValueChanged struct) and there is no need to keep them on the original struct if they are unused. They are being updated, but the logic is flawed as we are passing the Circuit as a value and not as a reference, therefore the update logic doesn't change the stored value.