Extend the HTTP Client to have an explicit API.
This change finalizes the previous work on providing explicit structs for all the objects returned by DS API and to provide functions to call the API directly from the client. This provides better abstraction as the underlying endpoints and parameters structure could be change now under the Client (e.g. migrating to new API version) without the need to change the modules that want to retrieve these endpoints.